backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
psycopg[binary]==3.9.19
python-dotenv==1.0.0
pydantic==2.6.4
pydantic-settings==2.2.1
qdrant-client==1.9.1
openai==1.48.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
httpx==0.27.2
cors==1.0.1
python-multipart==0.0.6
PyJWT==2.10.0
cryptography==42.0.7
